+++ lib/connection.py
@@ -2,20 +2,20 @@
 #########################################################################
 #  Copyright 2013 Marcus Popp                              marcus@popp.mx
 #########################################################################
-#  This file is part of SmartHome.py.    http://mknx.github.io/smarthome/
+#  This file is part of SmartHomeNG.    https://github.com/smarthomeNG//
 #
-#  SmartHome.py is free software: you can redistribute it and/or modify
+#  SmartHomeNG is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
 #  (at your option) any later version.
 #
-#  SmartHome.py is distributed in the hope that it will be useful,
+#  SmartHomeNG is distributed in the hope that it will be useful,
 #  but WITHOUT ANY WARRANTY; without even the implied warranty of
 #  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #  GNU General Public License for more details.
 #
 #  You should have received a copy of the GNU General Public License
-#  along with SmartHome.py. If not, see <http://www.gnu.org/licenses/>.
+#  along with SmartHomeNG. If not, see <http://www.gnu.org/licenses/>.
 #########################################################################
 
 import logging
@@ -288,9 +288,6 @@ class Stream(Base):
                 sent = self.socket.send(frame)
                 if sent < len(frame):
                     self.outbuffer.append(frame[sent:])
-            if self._close_after_send:
-                logger.debug("close after send")
-                self.close()
         except IndexError:  # buffer empty
             return
         except socket.error:
@@ -299,6 +296,9 @@ class Stream(Base):
             logger.exception("{}: {}".format(self._name, e))
             self.close()
         finally:
+            if self._close_after_send:
+                logger.debug("close after send")
+                self.close()
             self.__olock.release()
 
     def balance(self, bopen, bclose):